% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-info.R
\name{external_info}
\alias{external_info}
\title{Information about related software}
\usage{
external_info()
}
\value{
A list with elements:
\itemize{
\item \code{cairo}: The cairo version string.
\item \code{libpng}: The png version string.
\item \code{jpeg}: The jpeg version string.
\item \code{tiff}: The tiff library and version string used.
\item \code{tcl}: The tcl version string.
\item \code{curl}: The curl version string.
\item \code{zlib}: The zlib version string.
\item \code{bzlib}: The zlib version string.
\item \code{xz}: The zlib version string.
\item \code{PCRE}: The Perl Compatible Regular Expressions (PCRE) version string.
\item \code{ICU}: The International Components for Unicode (ICU) version string.
\item \code{TRE}: The TRE version string.
\item \code{iconv}: The iconv version string.
\item \code{readline}: The readline version string.
\item \code{BLAS}: The path with the implementation of BLAS in use.
\item \code{LAPACK}: The path with the implementation of LAPACK in use.
}
}
\description{
Information about related software
}
\details{
Note that calling this function will attempt to load the tcltk and
grDevices packages.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
external_info()
\dontshow{\}) # examplesIf}
}
\seealso{
Similar functions and objects in the base packages:
\code{\link[utils:sessionInfo]{utils::sessionInfo()}}, \link[base:extSoftVersion]{base::extSoftVersion}, \code{\link[tcltk:tclVersion]{tcltk::tclVersion()}}
\link[base:La_library]{base::La_library}, \code{\link[base:La_version]{base::La_version()}}, \code{\link[base:libcurlVersion]{base::libcurlVersion()}}.
}
