% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osname.R
\name{os_name}
\alias{os_name}
\title{Human readable name of the current operating system}
\usage{
os_name()
}
\value{
A character scalar.
}
\description{
For example Windows 8.1 instead of Windows version 6.3.9600.
On macOS it includes the code names, on Linux it includes the
distribution names and codenames if appropriate.
}
\details{
It uses \code{\link[utils:sessionInfo]{utils::sessionInfo()}}, but simplifies its output a bit
on Windows, to make it more concise.
}
