% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ sessioninfo }
\alias{ sessioninfo }
\alias{ sessioninfo }
\title{ R Session Information }
\description{
Query and print information about the current R session. It is similar to
  'utils::sessionInfo()`, but includes more information about packages, and
  where they were installed from.
}
