% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform-info.R
\name{platform_info}
\alias{platform_info}
\title{Information about the current platform}
\usage{
platform_info()
}
\value{
A list with elements:
  * `version`: the R version string.
  * `os`: the OS name in human readable format, see [osname()].
  * `system`: CPU, and machine readable OS name, separated by a comma.
  * `ui`: the user interface, e.g. `Rgui`, `RTerm`, etc. see `GUI`
    in [base::.Platform].
  * `language`: The current language setting. The `LANGUAGE` environment
    variable, if set, or `(EN)` if unset.
  * `collate`: Collation rule, from the current locale.
  * `tz`: The current time zone.
  * `date`: The current date.
}
\description{
Information about the current platform
}
\examples{
platform_info()
}
\seealso{
Similar functions and objects in the base packages:
  [base::R.version.string], [utils::sessionInfo()], [base::version],
  [base::.Platform], [base::Sys.getlocale()], [base::Sys.timezone()].
}
