% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessioninfo-package.R
\docType{package}
\name{sessioninfo-package}
\alias{sessioninfo}
\alias{sessioninfo-package}
\title{sessioninfo: R Session Information}
\description{
Query and print information about the current R session. It is similar to 'utils::sessionInfo()', but includes more information about packages, and where they were installed from.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/sessioninfo#readme}
  \item \url{https://sessioninfo.r-lib.org}
  \item Report bugs at \url{https://github.com/r-lib/sessioninfo/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Hadley Wickham
  \item Winston Chang
  \item Robert Flight
  \item Kirill Müller
  \item Jim Hester
}

Other contributors:
\itemize{
  \item R Core team [contributor]
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
